Remove deprecated model versions
