--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0/metadata-must-be-object.errors
@@ -1,1 +1,0 @@
-[ERROR] -: `metadata` must be an object | Model
