--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0/metadata-must-be-object.json
@@ -1,4 +1,0 @@
-{
-    "smithy": "0.4.0",
-    "metadata": "foo"
-}
