--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0/trait-duplicates.errors
@@ -1,1 +1,0 @@
-[ERROR] ns.foo#Invalid1: Conflicting `smithy.api#documentation` trait found on shape `ns.foo#Invalid1`. | Model
