--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0/trait-duplicates.json
@@ -1,34 +1,0 @@
-{
-    "smithy": "0.4.0",
-    "ns.foo": {
-        "shapes": {
-            "Valid1": {
-                "type": "string",
-                "sensitive": true
-            },
-            "Valid2": {
-                "type": "string",
-                "tags": [
-                    "foo"
-                ]
-            },
-            "Invalid1": {
-                "type": "string",
-                "documentation": "Abc"
-            }
-        },
-        "traits": {
-            "Valid1": {
-                "sensitive": true
-            },
-            "Valid2": {
-                "tags": [
-                    "baz"
-                ]
-            },
-            "Invalid1": {
-                "documentation": "Def"
-            }
-        }
-    }
-}
