--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0/validators-and-suppressions-must-be-arrays.errors
@@ -1,2 +1,0 @@
-[ERROR] -: suppressions must be an array. Found object. | Model
-[ERROR] -: validators must be an array. Found object. | Model
