--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0/validators-and-suppressions-must-be-arrays.json
@@ -1,9 +1,0 @@
-{
-    "smithy": "0.4.0",
-    "metadata": {
-        "validators": {
-        },
-        "suppressions": {
-        }
-    }
-}
