--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0_compat/metadata-must-come-before-namespace.errors
@@ -1,1 +1,0 @@
-[WARNING] -: Detected deprecated IDL features that will break in future versions of Smithy: Metadata statements must appear before a namespace statement | Model
