--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0_compat/metadata-must-come-before-namespace.smithy
@@ -1,3 +1,0 @@
-namespace com.foo
-
-metadata foo = "hello"
