--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0_compat/one-namespace-per-file.errors
@@ -1,1 +1,0 @@
-[WARNING] -: Detected deprecated IDL features that will break in future versions of Smithy: Only a single namespace can be declared per/file. The previous namespace was set to `com.foo`. | Model
