--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0_compat/one-namespace-per-file.smithy
@@ -1,2 +1,0 @@
-namespace com.foo
-namespace com.baz
