--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0_compat/use-must-come-after-namespace.errors
@@ -1,1 +1,0 @@
-[WARNING] -: Detected deprecated IDL features that will break in future versions of Smithy: Use statements must appear after a namespace is defined | Model
