--- smithy-model/src/test/resources/software/amazon/smithy/model/errorfiles/loader/version_0_4_0_compat/use-must-come-after-namespace.smithy
@@ -1,5 +1,0 @@
-use smithy.example#Foo
-
-namespace com.foo
-
-string MyString
