+++ smithy-model/src/test/resources/software/amazon/smithy/model/loader/invalid/invalid-unquoted-shape-id.smithy
@@ -1,5 +1,5 @@
 // Error resolving shape target; Invalid shape ID: __$#
 
-namespace foo.baz
-
 metadata abc = __$#
+
+namespace foo.baz
