+++ smithy-model/src/test/resources/software/amazon/smithy/model/loader/invalid/use/conflicting-shape-def.smithy
@@ -1,6 +1,4 @@
 // shape name `String` conflicts with imported shape `smithy.api#String`
-namespace foo.baz
-
 namespace smithy.example
 
 use smithy.api#String
