+++ smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/metadata.json
@@ -72,6 +72,6 @@
             "a",
             "b"
         ],
-        "shape_id": "foo.baz#Bam"
+        "shape_id": "smithy.api#String"
     }
 }
