+++ smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/multiple-traits-same-name.smithy
@@ -1,52 +1,13 @@
 namespace foo.alpha
 
-@hello
-string MyString1
+@deprecated
+string DeprecatedString1
 
-@foo.alpha#hello
-string MyString2
-
-@foo.beta#hello
-string MyString3
-
-@foo.gamma#hello
-string MyString4
+@smithy.api#deprecated
+string DeprecatedString2
 
 @trait(selector: "*")
-structure hello {}
+structure deprecated {}
 
-
-namespace foo.beta
-
-@hello
-string MyString1
-
-@foo.alpha#hello
-string MyString2
-
-@foo.beta#hello
-string MyString3
-
-@foo.gamma#hello
-string MyString4
-
-@trait(selector: "*")
-structure hello {}
-
-
-namespace foo.gamma
-
-@hello
-string MyString1
-
-@foo.alpha#hello
-string MyString2
-
-@foo.beta#hello
-string MyString3
-
-@foo.gamma#hello
-string MyString4
-
-@trait(selector: "*")
-structure hello {}
+@deprecated
+string DeprecatedString3
