--- smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/version_0_4_0/metadata-can-come-after-namespace.json
@@ -1,7 +1,0 @@
-{
-    "smithy": "0.5.0",
-    "metadata": {
-        "foo": "hello"
-    },
-    "shapes": { }
-}
