--- smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/version_0_4_0/metadata-can-come-after-namespace.smithy
@@ -1,4 +1,0 @@
-$version: "0.4.0"
-namespace com.foo
-
-metadata foo = "hello"
