--- smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/version_0_4_0/multiple-namespaces-per-file.smithy
@@ -1,5 +1,0 @@
-$version: "0.4.0"
-namespace com.foo
-namespace com.baz
-
-string MyString
