--- smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/version_0_4_0/use-before-namespace.json
@@ -1,8 +1,0 @@
-{
-    "smithy": "0.5.0",
-    "shapes": {
-        "com.baz#MyString": {
-            "type": "string"
-        }
-    }
-}
