--- smithy-model/src/test/resources/software/amazon/smithy/model/loader/valid/version_0_4_0/use-before-namespace.smithy
@@ -1,6 +1,0 @@
-$version: "0.4.0"
-use smithy.example#Foo
-
-namespace com.baz
-
-string MyString
